refactor(tests): move v3 meters tests to property views
