+++ seed/tests/test_import_file_views.py
@@ -3,9 +3,6 @@
 
 import ast
 import os
-import json
-
-from config.settings.common import TIME_ZONE
 
 from datetime import datetime
 
@@ -13,18 +10,12 @@ from django.core.files.uploadedfile import SimpleUploadedFile
 from django.urls import reverse
 from django.utils.timezone import (
     get_current_timezone,
-    make_aware,
 )
 
-from pytz import timezone
-
 from seed.data_importer.models import ImportFile, ImportRecord
-from seed.data_importer.tasks import save_raw_data
 from seed.landing.models import SEEDUser as User
 from seed.lib.superperms.orgs.models import Organization
 from seed.models import (
-    Meter,
-    MeterReading,
     PropertyState,
     PropertyView,
 )
@@ -337,626 +328,3 @@ class TestMeterViewSet(DataMappingBaseTestCase):
         ]
 
         self.assertCountEqual(result_dict.get("unlinkable_pm_ids"), expectation)
-
-    def test_property_meters_endpoint_returns_a_list_of_meters_of_a_view(self):
-        # add meters and readings to property associated to property_view_1
-        save_raw_data(self.import_file.id)
-
-        # create GB gas meter
-        meter_details = {
-            'source': Meter.GREENBUTTON,
-            'source_id': '/v1/User/000/UsagePoint/123fakeID/MeterReading/000',
-            'type': Meter.NATURAL_GAS,
-            'property_id': self.property_view_1.property.id,
-        }
-        gb_gas_meter = Meter.objects.create(**meter_details)
-
-        url = reverse('api:v2:meters-property-meters')
-
-        post_params = json.dumps({
-            'property_view_id': self.property_view_1.id,
-        })
-
-        result = self.client.post(url, post_params, content_type="application/json")
-        result_dict = json.loads(result.content)
-
-        electric_meter = Meter.objects.get(property_id=self.property_view_1.property_id, type=Meter.ELECTRICITY_GRID)
-        gas_meter = Meter.objects.get(property_id=self.property_view_1.property_id, type=Meter.NATURAL_GAS, source=Meter.PORTFOLIO_MANAGER)
-        expectation = [
-            {
-                'id': electric_meter.id,
-                'type': 'Electric - Grid',
-                'source': 'PM',
-                'source_id': '5766973-0',
-                'scenario_id': None,
-                'scenario_name': None
-            }, {
-                'id': gas_meter.id,
-                'type': 'Natural Gas',
-                'source': 'PM',
-                'source_id': '5766973-1',
-                'scenario_id': None,
-                'scenario_name': None
-            }, {
-                'id': gb_gas_meter.id,
-                'type': 'Natural Gas',
-                'source': 'GB',
-                'source_id': '123fakeID',
-                'scenario_id': None,
-                'scenario_name': None
-            },
-        ]
-
-        self.assertCountEqual(result_dict, expectation)
-
-    def test_property_meter_usage_returns_meter_readings_and_column_defs_given_property_view_and_nondefault_meter_display_org_settings(self):
-        # Update settings for display meter units to change it from the default values.
-        self.org.display_meter_units['Electric - Grid'] = 'kWh (thousand Watt-hours)'
-        self.org.display_meter_units['Natural Gas'] = 'kcf (thousand cubic feet)'
-        self.org.save()
-
-        # add meters and readings to property associated to property_view_1
-        save_raw_data(self.import_file.id)
-
-        meter_details = {
-            'source': Meter.GREENBUTTON,
-            'source_id': '/v1/User/000/UsagePoint/123fakeID/MeterReading/000',
-            'type': Meter.NATURAL_GAS,
-            'property_id': self.property_view_1.property.id,
-        }
-        gb_gas_meter = Meter.objects.create(**meter_details)
-
-        tz_obj = timezone(TIME_ZONE)
-        gb_gas_reading_details = {
-            'start_time': make_aware(datetime(2016, 1, 1, 0, 0, 0), timezone=tz_obj),
-            'end_time': make_aware(datetime(2016, 2, 1, 0, 0, 0), timezone=tz_obj),
-            'reading': 1000,
-            'source_unit': 'kBtu (thousand Btu)',
-            'conversion_factor': 1,
-            'meter_id': gb_gas_meter.id,
-        }
-        MeterReading.objects.create(**gb_gas_reading_details)
-
-        url = reverse('api:v2:meters-property-meter-usage')
-
-        post_params = json.dumps({
-            'property_view_id': self.property_view_1.id,
-            'interval': 'Exact',
-            'excluded_meter_ids': [],
-        })
-        result = self.client.post(url, post_params, content_type="application/json")
-        result_dict = ast.literal_eval(result.content.decode("utf-8"))
-
-        expectation = {
-            'readings': [
-                {
-                    'start_time': '2016-01-01 00:00:00',
-                    'end_time': '2016-02-01 00:00:00',
-                    'Electric - Grid - PM - 5766973-0': (597478.9 / 3.41),
-                    'Natural Gas - PM - 5766973-1': 576000.2 / 1026,
-                    'Natural Gas - GB - 123fakeID': 1000 / 1026,
-                },
-                {
-                    'start_time': '2016-02-01 00:00:00',
-                    'end_time': '2016-03-01 00:00:00',
-                    'Electric - Grid - PM - 5766973-0': (548603.7 / 3.41),
-                    'Natural Gas - PM - 5766973-1': 488000.1 / 1026,
-                },
-            ],
-            'column_defs': [
-                {
-                    'field': 'start_time',
-                    '_filter_type': 'datetime',
-                },
-                {
-                    'field': 'end_time',
-                    '_filter_type': 'datetime',
-                },
-                {
-                    'field': 'Electric - Grid - PM - 5766973-0',
-                    'displayName': 'Electric - Grid - PM - 5766973-0 (kWh (thousand Watt-hours))',
-                    '_filter_type': 'reading',
-                },
-                {
-                    'field': 'Natural Gas - PM - 5766973-1',
-                    'displayName': 'Natural Gas - PM - 5766973-1 (kcf (thousand cubic feet))',
-                    '_filter_type': 'reading',
-                },
-                {
-                    'field': 'Natural Gas - GB - 123fakeID',
-                    'displayName': 'Natural Gas - GB - 123fakeID (kcf (thousand cubic feet))',
-                    '_filter_type': 'reading',
-                },
-            ]
-        }
-
-        self.assertCountEqual(result_dict['readings'], expectation['readings'])
-        self.assertCountEqual(result_dict['column_defs'], expectation['column_defs'])
-
-    def test_property_meter_usage_returns_meter_readings_and_column_defs_when_cost_meter_included(self):
-        filename = "example-pm-monthly-meter-usage-2-cost-meters.xlsx"
-        filepath = os.path.dirname(os.path.abspath(__file__)) + "/data/" + filename
-
-        cost_import_file = ImportFile.objects.create(
-            import_record=self.import_record,
-            source_type="PM Meter Usage",
-            uploaded_filename=filename,
-            file=SimpleUploadedFile(name=filename, content=open(filepath, 'rb').read()),
-            cycle=self.cycle
-        )
-
-        # add meters and readings to property associated to property_view_1
-        save_raw_data(cost_import_file.id)
-
-        url = reverse('api:v2:meters-property-meter-usage')
-
-        post_params = json.dumps({
-            'property_view_id': self.property_view_1.id,
-            'interval': 'Exact',
-            'excluded_meter_ids': [],
-        })
-        result = self.client.post(url, post_params, content_type="application/json")
-        result_dict = ast.literal_eval(result.content.decode("utf-8"))
-
-        expectation = {
-            'readings': [
-                {
-                    'start_time': '2016-01-01 00:00:00',
-                    'end_time': '2016-02-01 00:00:00',
-                    'Electric - Grid - PM - 5766973-0': 597478.9 / 3.41,
-                    'Cost - PM - 5766973-0': 100,
-                    'Natural Gas - PM - 5766973-1': 576000.2,
-                    'Cost - PM - 5766973-1': 300,
-                },
-                {
-                    'start_time': '2016-02-01 00:00:00',
-                    'end_time': '2016-03-01 00:00:00',
-                    'Electric - Grid - PM - 5766973-0': 548603.7 / 3.41,
-                    'Cost - PM - 5766973-0': 200,
-                    'Natural Gas - PM - 5766973-1': 488000.1,
-                    'Cost - PM - 5766973-1': 400,
-                },
-            ],
-            'column_defs': [
-                {
-                    'field': 'start_time',
-                    '_filter_type': 'datetime',
-                },
-                {
-                    'field': 'end_time',
-                    '_filter_type': 'datetime',
-                },
-                {
-                    'field': 'Electric - Grid - PM - 5766973-0',
-                    'displayName': 'Electric - Grid - PM - 5766973-0 (kWh (thousand Watt-hours))',
-                    '_filter_type': 'reading',
-                },
-                {
-                    'field': 'Natural Gas - PM - 5766973-1',
-                    'displayName': 'Natural Gas - PM - 5766973-1 (kBtu (thousand Btu))',
-                    '_filter_type': 'reading',
-                },
-                {
-                    'field': 'Cost - PM - 5766973-0',
-                    'displayName': 'Cost - PM - 5766973-0 (US Dollars)',
-                    '_filter_type': 'reading',
-                },
-                {
-                    'field': 'Cost - PM - 5766973-1',
-                    'displayName': 'Cost - PM - 5766973-1 (US Dollars)',
-                    '_filter_type': 'reading',
-                },
-            ]
-        }
-
-        self.assertCountEqual(result_dict['readings'], expectation['readings'])
-        self.assertCountEqual(result_dict['column_defs'], expectation['column_defs'])
-
-    def test_property_meter_usage_returns_meter_readings_according_to_thermal_conversion_preferences_of_an_org_if_applicable_for_display_settings(self):
-        # update the org settings thermal preference and display preference
-        self.org.thermal_conversion_assumption = Organization.CAN
-        self.org.display_meter_units["Diesel"] = "Liters"
-        self.org.display_meter_units["Coke"] = "Lbs. (pounds)"
-        self.org.save()
-
-        # add meters and readings to property associated to property_view_1
-        meter_details = {
-            'source': Meter.PORTFOLIO_MANAGER,
-            'source_id': '123fakeID',
-            'type': Meter.DIESEL,
-            'property_id': self.property_view_1.property.id,
-        }
-        diesel_meter = Meter.objects.create(**meter_details)
-
-        tz_obj = timezone(TIME_ZONE)
-        diesel_reading_details = {
-            'start_time': make_aware(datetime(2016, 1, 1, 0, 0, 0), timezone=tz_obj),
-            'end_time': make_aware(datetime(2016, 2, 1, 0, 0, 0), timezone=tz_obj),
-            'reading': 10,
-            'source_unit': 'kBtu (thousand Btu)',
-            'conversion_factor': 1,
-            'meter_id': diesel_meter.id,
-        }
-        MeterReading.objects.create(**diesel_reading_details)
-
-        meter_details['type'] = Meter.COKE
-        meter_details['source_id'] = '456fakeID'
-        coke_meter = Meter.objects.create(**meter_details)
-
-        coke_reading_details = {
-            'start_time': make_aware(datetime(2016, 1, 1, 0, 0, 0), timezone=tz_obj),
-            'end_time': make_aware(datetime(2016, 2, 1, 0, 0, 0), timezone=tz_obj),
-            'reading': 100,
-            'source_unit': 'kBtu (thousand Btu)',
-            'conversion_factor': 1,
-            'meter_id': coke_meter.id,
-        }
-        MeterReading.objects.create(**coke_reading_details)
-
-        post_params = json.dumps({
-            'property_view_id': self.property_view_1.id,
-            'interval': 'Exact',
-            'excluded_meter_ids': [],
-        })
-
-        url = reverse('api:v2:meters-property-meter-usage')
-        result = self.client.post(url, post_params, content_type="application/json")
-        result_dict = ast.literal_eval(result.content.decode("utf-8"))
-
-        display_readings = [
-            {
-                'start_time': '2016-01-01 00:00:00',
-                'end_time': '2016-02-01 00:00:00',
-                'Diesel - PM - 123fakeID': 10 / 36.30,
-                'Coke - PM - 456fakeID': 100 / 12.39,
-            },
-        ]
-
-        self.assertCountEqual(result_dict['readings'], display_readings)
-
-    def test_property_meter_usage_can_return_monthly_meter_readings_and_column_defs_with_nondefault_display_setting(self):
-        # Update settings for display meter units to change it from the default values.
-        self.org.display_meter_units['Electric - Grid'] = 'kWh (thousand Watt-hours)'
-        self.org.save()
-
-        # add initial meters and readings
-        save_raw_data(self.import_file.id)
-
-        # add additional entries for each initial meter
-        tz_obj = timezone(TIME_ZONE)
-        for meter in Meter.objects.all():
-            # March 2016 reading
-            reading_details = {
-                'meter_id': meter.id,
-                'start_time': make_aware(datetime(2016, 3, 1, 0, 0, 0), timezone=tz_obj),
-                'end_time': make_aware(datetime(2016, 4, 1, 0, 0, 0), timezone=tz_obj),
-                'reading': 100,
-                'source_unit': 'kBtu (thousand Btu)',
-                'conversion_factor': 1
-            }
-            MeterReading.objects.create(**reading_details)
-
-            # May 2016 reading
-            reading_details['start_time'] = make_aware(datetime(2016, 5, 1, 0, 0, 0), timezone=tz_obj)
-            reading_details['end_time'] = make_aware(datetime(2016, 6, 1, 0, 0, 0), timezone=tz_obj)
-            reading_details['reading'] = 200
-            MeterReading.objects.create(**reading_details)
-
-        url = reverse('api:v2:meters-property-meter-usage')
-
-        post_params = json.dumps({
-            'property_view_id': self.property_view_1.id,
-            'interval': 'Month',
-            'excluded_meter_ids': [],
-        })
-        result = self.client.post(url, post_params, content_type="application/json")
-        result_dict = ast.literal_eval(result.content.decode("utf-8"))
-
-        expectation = {
-            'readings': [
-                {
-                    'month': 'January 2016',
-                    'Electric - Grid - PM - 5766973-0': 597478.9 / 3.41,
-                    'Natural Gas - PM - 5766973-1': 576000.2,
-                },
-                {
-                    'month': 'February 2016',
-                    'Electric - Grid - PM - 5766973-0': 548603.7 / 3.41,
-                    'Natural Gas - PM - 5766973-1': 488000.1,
-                },
-                {
-                    'month': 'March 2016',
-                    'Electric - Grid - PM - 5766973-0': 100 / 3.41,
-                    'Natural Gas - PM - 5766973-1': 100,
-                },
-                {
-                    'month': 'May 2016',
-                    'Electric - Grid - PM - 5766973-0': 200 / 3.41,
-                    'Natural Gas - PM - 5766973-1': 200,
-                },
-            ],
-            'column_defs': [
-                {
-                    'field': 'month',
-                    '_filter_type': 'datetime',
-                },
-                {
-                    'field': 'Electric - Grid - PM - 5766973-0',
-                    'displayName': 'Electric - Grid - PM - 5766973-0 (kWh (thousand Watt-hours))',
-                    '_filter_type': 'reading',
-                },
-                {
-                    'field': 'Natural Gas - PM - 5766973-1',
-                    'displayName': 'Natural Gas - PM - 5766973-1 (kBtu (thousand Btu))',
-                    '_filter_type': 'reading',
-                },
-            ]
-        }
-
-        self.assertCountEqual(result_dict['readings'], expectation['readings'])
-        self.assertCountEqual(result_dict['column_defs'], expectation['column_defs'])
-
-    def test_property_meter_usage_can_return_monthly_meter_readings_and_column_defs_for_submonthly_data_with_DST_transitions_and_specific_meters(self):
-        # add initial meters and readings
-        save_raw_data(self.import_file.id)
-
-        property_1_electric_meter = Meter.objects.get(source_id='5766973-0')
-        # add additional sub-montly entries for each initial meter
-        tz_obj = timezone(TIME_ZONE)
-        for meter in Meter.objects.all():
-            # November 2019 reading between DST transition
-            reading_details = {
-                'meter_id': meter.id,
-                'start_time': make_aware(datetime(2019, 11, 3, 1, 59, 59), timezone=tz_obj, is_dst=True),
-                'end_time': make_aware(datetime(2019, 11, 3, 1, 59, 59), timezone=tz_obj, is_dst=False),
-                'reading': 100,
-                'source_unit': 'kBtu (thousand Btu)',
-                'conversion_factor': 1
-            }
-            MeterReading.objects.create(**reading_details)
-
-            # November 2019 reading after DST transition
-            reading_details['start_time'] = make_aware(datetime(2019, 11, 3, 2, 0, 0), timezone=tz_obj)
-            reading_details['end_time'] = make_aware(datetime(2019, 11, 3, 3, 0, 0), timezone=tz_obj)
-            reading_details['reading'] = 200
-            MeterReading.objects.create(**reading_details)
-
-            # Create a reading for only one of the meters that will be filtered out completely
-            if meter.source_id == property_1_electric_meter.id:
-                reading_details['start_time'] = make_aware(datetime(2020, 11, 3, 2, 0, 0), timezone=tz_obj)
-                reading_details['end_time'] = make_aware(datetime(2020, 11, 3, 3, 0, 0), timezone=tz_obj)
-                reading_details['reading'] = 10000000
-                MeterReading.objects.create(**reading_details)
-
-        url = reverse('api:v2:meters-property-meter-usage')
-
-        post_params = json.dumps({
-            'property_view_id': self.property_view_1.id,
-            'interval': 'Month',
-            'excluded_meter_ids': [property_1_electric_meter.id],
-        })
-        result = self.client.post(url, post_params, content_type="application/json")
-        result_dict = ast.literal_eval(result.content.decode("utf-8"))
-
-        expectation = {
-            'readings': [
-                {
-                    'month': 'January 2016',
-                    'Natural Gas - PM - 5766973-1': 576000.2,
-                },
-                {
-                    'month': 'February 2016',
-                    'Natural Gas - PM - 5766973-1': 488000.1,
-                },
-                {
-                    'month': 'November 2019',
-                    'Natural Gas - PM - 5766973-1': 300,
-                },
-            ],
-            'column_defs': [
-                {
-                    'field': 'month',
-                    '_filter_type': 'datetime',
-                },
-                {
-                    'field': 'Natural Gas - PM - 5766973-1',
-                    'displayName': 'Natural Gas - PM - 5766973-1 (kBtu (thousand Btu))',
-                    '_filter_type': 'reading',
-                },
-            ]
-        }
-
-        self.assertCountEqual(result_dict['readings'], expectation['readings'])
-        self.assertCountEqual(result_dict['column_defs'], expectation['column_defs'])
-
-    def test_property_meter_usage_can_return_monthly_meter_readings_and_column_defs_of_overlapping_submonthly_data_aggregating_monthly_data_to_maximize_total(self):
-        # add initial meters and readings
-        save_raw_data(self.import_file.id)
-
-        # add additional entries for the Electricity meter
-        tz_obj = timezone(TIME_ZONE)
-        meter = Meter.objects.get(property_id=self.property_view_1.property.id, type=Meter.type_lookup['Electric - Grid'])
-        # 2016 January reading that should override the existing reading
-        reading_details = {
-            'meter_id': meter.id,
-            'start_time': make_aware(datetime(2016, 1, 1, 0, 0, 0), timezone=tz_obj),
-            'end_time': make_aware(datetime(2016, 1, 20, 23, 59, 59), timezone=tz_obj),
-            'reading': 100000000000000,
-            'source_unit': 'kBtu (thousand Btu)',
-            'conversion_factor': 1
-        }
-        MeterReading.objects.create(**reading_details)
-
-        # 2016 January reading that should be ignored
-        reading_details['start_time'] = make_aware(datetime(2016, 1, 1, 0, 0, 0), timezone=tz_obj)
-        reading_details['end_time'] = make_aware(datetime(2016, 3, 31, 23, 59, 59), timezone=tz_obj)
-        reading_details['reading'] = 0.1
-        MeterReading.objects.create(**reading_details)
-
-        # Create March 2016 entries having disregarded readings when finding monthly total
-        # 1 week - not included in total
-        reading_details['start_time'] = make_aware(datetime(2016, 3, 1, 0, 0, 0), timezone=tz_obj)
-        reading_details['end_time'] = make_aware(datetime(2016, 3, 6, 23, 59, 59), timezone=tz_obj)
-        reading_details['reading'] = 1
-        MeterReading.objects.create(**reading_details)
-
-        # 1 week - not included in total
-        reading_details['start_time'] = make_aware(datetime(2016, 3, 7, 0, 0, 0), timezone=tz_obj)
-        reading_details['end_time'] = make_aware(datetime(2016, 3, 13, 23, 59, 59), timezone=tz_obj)
-        reading_details['reading'] = 10
-        MeterReading.objects.create(**reading_details)
-
-        # 10 days - included in total
-        reading_details['start_time'] = make_aware(datetime(2016, 3, 2, 0, 0, 0), timezone=tz_obj)
-        reading_details['end_time'] = make_aware(datetime(2016, 3, 11, 23, 59, 59), timezone=tz_obj)
-        reading_details['reading'] = 100
-        MeterReading.objects.create(**reading_details)
-
-        # 10 days - included in total
-        reading_details['start_time'] = make_aware(datetime(2016, 3, 12, 0, 0, 0), timezone=tz_obj)
-        reading_details['end_time'] = make_aware(datetime(2016, 3, 21, 23, 59, 59), timezone=tz_obj)
-        reading_details['reading'] = 1000
-        MeterReading.objects.create(**reading_details)
-
-        # Create April 2016 entries having disregarded readings when finding monthly total
-        # 5 days - not included in total
-        reading_details['start_time'] = make_aware(datetime(2016, 4, 1, 0, 0, 0), timezone=tz_obj)
-        reading_details['end_time'] = make_aware(datetime(2016, 4, 4, 23, 59, 59), timezone=tz_obj)
-        reading_details['reading'] = 2
-        MeterReading.objects.create(**reading_details)
-
-        # 10 days - not included in total
-        reading_details['start_time'] = make_aware(datetime(2016, 4, 6, 0, 0, 0), timezone=tz_obj)
-        reading_details['end_time'] = make_aware(datetime(2016, 4, 15, 23, 59, 59), timezone=tz_obj)
-        reading_details['reading'] = 20
-        MeterReading.objects.create(**reading_details)
-
-        # 20 days - included in total
-        reading_details['start_time'] = make_aware(datetime(2016, 4, 2, 0, 0, 0), timezone=tz_obj)
-        reading_details['end_time'] = make_aware(datetime(2016, 4, 21, 23, 59, 59), timezone=tz_obj)
-        reading_details['reading'] = 200
-        MeterReading.objects.create(**reading_details)
-
-        url = reverse('api:v2:meters-property-meter-usage')
-
-        post_params = json.dumps({
-            'property_view_id': self.property_view_1.id,
-            'interval': 'Month',
-            'excluded_meter_ids': [],
-        })
-        result = self.client.post(url, post_params, content_type="application/json")
-        result_dict = ast.literal_eval(result.content.decode("utf-8"))
-
-        expectation = {
-            'readings': [
-                {
-                    'month': 'January 2016',
-                    'Electric - Grid - PM - 5766973-0': 100000000000000 / 3.41,
-                    'Natural Gas - PM - 5766973-1': 576000.2,
-                },
-                {
-                    'month': 'February 2016',
-                    'Electric - Grid - PM - 5766973-0': 548603.7 / 3.41,
-                    'Natural Gas - PM - 5766973-1': 488000.1,
-                },
-                {
-                    'month': 'March 2016',
-                    'Electric - Grid - PM - 5766973-0': 1100 / 3.41,
-                },
-                {
-                    'month': 'April 2016',
-                    'Electric - Grid - PM - 5766973-0': 200 / 3.41,
-                },
-            ],
-            'column_defs': [
-                {
-                    'field': 'month',
-                    '_filter_type': 'datetime',
-                },
-                {
-                    'field': 'Electric - Grid - PM - 5766973-0',
-                    'displayName': 'Electric - Grid - PM - 5766973-0 (kWh (thousand Watt-hours))',
-                    '_filter_type': 'reading',
-                },
-                {
-                    'field': 'Natural Gas - PM - 5766973-1',
-                    'displayName': 'Natural Gas - PM - 5766973-1 (kBtu (thousand Btu))',
-                    '_filter_type': 'reading',
-                },
-            ]
-        }
-
-        self.assertCountEqual(result_dict['readings'], expectation['readings'])
-        self.assertCountEqual(result_dict['column_defs'], expectation['column_defs'])
-
-    def test_property_meter_usage_can_return_annual_meter_readings_and_column_defs_while_handling_a_nondefault_display_setting(self):
-        # Update settings for display meter units to change it from the default values.
-        self.org.display_meter_units['Electric - Grid'] = 'kWh (thousand Watt-hours)'
-        self.org.save()
-
-        # add initial meters and readings
-        save_raw_data(self.import_file.id)
-
-        # add additional 2018 entries for each initial meter
-        tz_obj = timezone(TIME_ZONE)
-        for meter in Meter.objects.all():
-            # March 2018 reading
-            reading_details = {
-                'meter_id': meter.id,
-                'start_time': make_aware(datetime(2018, 3, 1, 0, 0, 0), timezone=tz_obj),
-                'end_time': make_aware(datetime(2018, 4, 1, 0, 0, 0), timezone=tz_obj),
-                'reading': 100,
-                'source_unit': 'kBtu (thousand Btu)',
-                'conversion_factor': 1
-            }
-            MeterReading.objects.create(**reading_details)
-
-            # May 2018 reading
-            reading_details['start_time'] = make_aware(datetime(2018, 5, 1, 0, 0, 0), timezone=tz_obj)
-            reading_details['end_time'] = make_aware(datetime(2018, 6, 1, 0, 0, 0), timezone=tz_obj)
-            reading_details['reading'] = 200
-            MeterReading.objects.create(**reading_details)
-
-        url = reverse('api:v2:meters-property-meter-usage')
-
-        post_params = json.dumps({
-            'property_view_id': self.property_view_1.id,
-            'interval': 'Year',
-            'excluded_meter_ids': [],
-        })
-        result = self.client.post(url, post_params, content_type="application/json")
-        result_dict = ast.literal_eval(result.content.decode("utf-8"))
-
-        expectation = {
-            'readings': [
-                {
-                    'year': 2016,
-                    'Electric - Grid - PM - 5766973-0': (597478.9 + 548603.7) / 3.41,
-                    'Natural Gas - PM - 5766973-1': 576000.2 + 488000.1,
-                },
-                {
-                    'year': 2018,
-                    'Electric - Grid - PM - 5766973-0': (100 + 200) / 3.41,
-                    'Natural Gas - PM - 5766973-1': 100 + 200,
-                },
-            ],
-            'column_defs': [
-                {
-                    'field': 'year',
-                    '_filter_type': 'datetime',
-                },
-                {
-                    'field': 'Electric - Grid - PM - 5766973-0',
-                    'displayName': 'Electric - Grid - PM - 5766973-0 (kWh (thousand Watt-hours))',
-                    '_filter_type': 'reading',
-                },
-                {
-                    'field': 'Natural Gas - PM - 5766973-1',
-                    'displayName': 'Natural Gas - PM - 5766973-1 (kBtu (thousand Btu))',
-                    '_filter_type': 'reading',
-                },
-            ]
-        }
-
-        self.assertCountEqual(result_dict['readings'], expectation['readings'])
-        self.assertCountEqual(result_dict['column_defs'], expectation['column_defs'])
